Implement Pie chart with mock data
