+++ app/src/main/java/com/prj666_183a06/xbudget/PieChartFrag.java
@@ -16,45 +16,45 @@ import com.github.mikephil.charting.components.Legend;
 
 public class PieChartFrag extends ReportActivity {
 
-    public static Fragment newInstance(){
-        return new PieChartFrag();
-    }
-
-    private PieChart mChart;
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.frag_simple_pie, container, false);
-
-        mChart = v.findViewById(R.id.pieChart1);
-        mChart.getDescription().setEnabled(false);
-
-        Typeface tf = Typeface.createFromAsset(getActivity().getAssets(), "OpenSans-Light.ttf");
-
-        mChart.setCenterTextTypeface(tf);
-        mChart.setCenterText(generateCenterText());
-        mChart.setCenterTextSize(10f);
-        mChart.setCenterTextTypeface(tf);
-
-        mChart.setHoleRadius(45f);
-        mChart.setTransparentCircleRadius(50f);
-
-        Legend l = mChart.getLegend();
-        l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
-        l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
-        l.setOrientation(Legend.LegendOrientation.VERTICAL);
-        l.setDrawInside(false);
-
-        mChart.setData(generatePieData());
-
-        return v;
-    }
-
-    private SpannableString generateCenterText() {
-        SpannableString s = new SpannableString("Revenes\nQuaters 2015");
-        s.setSpan(new RelativeSizeSpan(2f), 0, 8, 0);
-        s.setSpan(new ForegroundColorSpan(Color.GRAY), 8, s.length(), 0);
-
-        return s;
-    }
+//    public static Fragment newInstance(){
+//        return new PieChartFrag();
+//    }
+//
+//    private PieChart mChart;
+//
+//    @Override
+//    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+//        View v = inflater.inflate(R.layout.frag_simple_pie, container, false);
+//
+//        mChart = v.findViewById(R.id.pieChart1);
+//        mChart.getDescription().setEnabled(false);
+//
+//        Typeface tf = Typeface.createFromAsset(getActivity().getAssets(), "OpenSans-Light.ttf");
+//
+//        mChart.setCenterTextTypeface(tf);
+//        mChart.setCenterText(generateCenterText());
+//        mChart.setCenterTextSize(10f);
+//        mChart.setCenterTextTypeface(tf);
+//
+//        mChart.setHoleRadius(45f);
+//        mChart.setTransparentCircleRadius(50f);
+//
+//        Legend l = mChart.getLegend();
+//        l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
+//        l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
+//        l.setOrientation(Legend.LegendOrientation.VERTICAL);
+//        l.setDrawInside(false);
+//
+//        mChart.setData(generatePieData());
+//
+//        return v;
+//    }
+//
+//    private SpannableString generateCenterText() {
+//        SpannableString s = new SpannableString("Revenes\nQuaters 2015");
+//        s.setSpan(new RelativeSizeSpan(2f), 0, 8, 0);
+//        s.setSpan(new ForegroundColorSpan(Color.GRAY), 8, s.length(), 0);
+//
+//        return s;
+//    }
 }
