+++ app/src/main/res/layout/activity_report.xml
@@ -5,17 +5,17 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent" >
 
-    <!--<TextView-->
-        <!--android:layout_centerHorizontal="true"-->
-        <!--android:layout_centerVertical="true"-->
-        <!--android:layout_width="wrap_content"-->
-        <!--android:layout_height="wrap_content"-->
-        <!--android:text="Report View"-->
-        <!--tools:layout_editor_absoluteX="162dp"-->
-        <!--tools:layout_editor_absoluteY="245dp" />-->
+    <TextView
+        android:layout_centerHorizontal="true"
+        android:layout_centerVertical="true"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Report View"
+        tools:layout_editor_absoluteX="162dp"
+        tools:layout_editor_absoluteY="245dp" />
 
     <com.github.mikephil.charting.charts.PieChart
-        android:id="@+id/chart1"
+        android:id="@+id/pieChart1"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
